1. ObjectPool Init이후에 사용하는지 확인할 수 있게 추가(Debug assert)
스트레스 봇에서 SessionBatchUpdater와 ThWorkerPool::Init 데이터 레이스 수정

+++ ServerCommon_LIB/Utility/private/Thread/ThreadPool.cpp
@@ -33,7 +33,7 @@ void ThreadPool::RunningThread(std::stop_token stopToken) {
       // Internal은 에러 코드
       // 64비트 자료형이지만, 실제로는 32비트만 사용
       // NTStatus에 대한 오류 코드임
-      workerJob(overlappedEntry[i].dwNumberOfBytesTransferred, overlappedEntry[i].Internal);
+      workerJob(overlappedEntry[i].dwNumberOfBytesTransferred, static_cast<DWORD>(overlappedEntry[i].Internal));
     }
     if (stopToken.stop_requested()) {
       return;
